added singleton support
