+++ lua/vimdaloo/lang/Object.lua
@@ -1,3 +1,22 @@
+-- https://www.lua.org/pil/19.3.html
+local function pairsByKeys(t, f)
+    local a = {}
+    for n in pairs(t) do
+        table.insert(a, n)
+    end
+    table.sort(a, f)
+    local i = 0 -- iterator variable
+    local iter = function() -- iterator function
+        i = i + 1
+        if a[i] == nil then
+            return nil
+        else
+            return a[i], t[a[i]]
+        end
+    end
+    return iter
+end
+
 namespace 'vimdaloo.lang' {
     ----
     --- &nbsp;<br/>
@@ -39,6 +58,31 @@ namespace 'vimdaloo.lang' {
                 self.class = vimdaloo.lang.Class(self.__getclass(self):getName())
             end
             return self.class
+        end,
+
+        --- string representation
+        -- @display toString
+        -- @treturn string the string representation
+        toString = function(self)
+            local str = ''
+            for key, val in pairsByKeys(self) do
+                if key ~= 'class' and string.sub(key, 1, 1) ~= '_' then
+                    if string.len(str) ~= 0 then
+                        str = str .. ','
+                    end
+                    str = string.format('%s%s=%s', str, key, val)
+                end
+            end
+            local classname = self.__getclass(self):getName()
+            if string.len(str) == 0 then
+                return string.format('%s', classname)
+            else
+                return string.format('%s(%s)', classname, str)
+            end
+        end,
+
+        __tostring = function(self)
+            return self:toString()
         end,
     },
 }
